Prompt 1
Can we add a User class that has an editable UserName and the ability to link to other platforms. Don't worry about platforms yet, just make the underlying code with a basic local only..  This should use Google Firebase APIs if needed for storing and logging in the user. I will add the private key later manually.