Prompt 2
That new UserProfileView.swift looks not within our current theme it's hard to separate everything,.  It also has a "Done" button, but the default Navigation Bar is showing the back button, first picture is showing these 2 issues.  The second picture shows another issue, in the ContentView.swift file, we have a second ">" on the new button to move to the Profile screen. The default one should be removed.  Please fix these.