Prompts 6,7 in USer branch
I want to add an avatar to the Users model.  It should be an enum, and have the following choices to begin with "red", "green", "blue", "orange", "purple", "yellow", "white", "black".  This should be set to a RANDOM color when a user is created if nothing is passed in.  Do not add it to the UserSettings window

Lets add avatatType, also randomly set.  Also an enum: woman, man, dog, cat, bird, car, building