Prompt 8 in User branch
Great, lets go back and fill in all of the FirebaseAuthService code as needed to support the following: have a default username that is uniquely created the first time the app is run on the phone . This will  be tied to the phone/appleid so it won't change unless they do manually.  They can then sign-in/link to Google, Apple, Android, Yahoo, or whatever else Firebase supports.  Once this is done, we will check the username for uniqueness in the list of user names.  If it has not been changed my the user, then we just generate a new one.  If the user has edited the name, we bring up a dialog to change the username, if they hit cancel, we stop and reverse  the linking/sign-in.

We need to add some User settings that show info we get from this linking, such as the email, phone number and any other identifiers.  We want them to choose if we make that public.  That way we can search for friends in a later update.

If we are already logged in and we change the user name, we should check to make sure that user name is available first and inform if not unique.