Prompts 9,10 in User Branch
Great, lets go back and fill in all of the FirebaseAuthService code as needed to support the following: have a default username that is uniquely created the first time the app is run on the phone . This will  be tied to the phone/appleid so it won't change unless they do manually.  They can then sign-in/link to Google, Apple, Android, Yahoo, or whatever else Firebase supports.  Once this is done, we will check the username for uniqueness in the list of user names.  If it has not been changed my the user, then we just generate a new one.  If the user has edited the name, we bring up a dialog to change the username, if they hit cancel, we stop and reverse  the linking/sign-in.

We need to add some User settings that show info we get from this linking, such as the email, phone number and any other identifiers.  We want them to choose if we make that public.  That way we can search for friends in a later update.

If we are already logged in and we change the user name, we should check to make sure that user name is available first and inform if not unique.

That works fine.  But you didn't move the toggle.  Could we instead maybe reuse DefaultSettingToggleRow or make a new one if it makes more sense?  We could add the hird detail row into that DefaultSettingsToggleRow.  While you are at it, ALWAYS make sure UI elements are resuable.  If it makes sense to make other Settings items reusable please do that now, with items like the username editable screen, or text only version, and the Profile button.