Prompt 12 in User Branch
Add a duplicate of SettingToggleRow combine it with the SettingsEditableTextRow, call it SettingShareDataToggleRow.  The only differences being you take in a value instead of the description and do the check to display "Not Set" in there, the description is the same .body size, not .caption.  It also should also take in an isEditable boolean, to determine if the pencil icon shows up and we can change it, as not all in the future will allow to be changed (default to false). Also move the Toggle itself to the bottom of the card using Spacer(), so the pencil is across from the valuve passed in.