Prompt 13 (and a wuestion) in User Branch.
Can you explain, don't change anything, why you went with onEdit() and pass a value, instead of onSave(), onCancel() and a binding?

Ok, that makes sense that if we have a Optional it would be a bit more complex. But I think I would rather separate the persistence.   Please refactor SettingShareDataToggleRow to be more like SettingEditableTextRow.  Make it the same in every way but having the toggle.

Manually made some design changes.