Prompts used for User Login.  Moved Assests file. Git ignore for Google Info files.  Some comments. Little if any manual code changes.
I want to go back and now fill in the FirebaseAuthService so that it connects to the Firebase Authorization Cloud service.  Make sure the app does not need internet for this app to work (has to work offline), including knowing who the User is and what they are logged into. Fix any conflict of UserID's when the internet is connected and the User info is pushed to the cloud service. We should allow them to sign-in using any service they want, or just standard username and email.  This should all be connected back to the @UserProfileView.swift so the user can sign in/create an account even if they are already playing with the default user.

The Apple Oauth is giving errors, line 629: Static member 'credential' cannot be used on instance of type 'OAuthProvider'

Still complaining, same error and even more now.

There seems to be no reference to FIREBASE_CONFIG_FILE.

When I try to create an account, I get this: 12.6.0 - [FirebaseFirestore][I-FST000001] WatchStream (313035623833383938) Stream error: 'Permission denied: Permission denied on resource project roadtrip-royale-dev.'

It seems the SignInView  that @UserProfileView.swift tries to access is not longer there. Maybe it was deleted, can you reproduce the same View?

I am still getting an error, slightly different: 12.6.0 - [FirebaseFirestore][I-FST000001] WatchStream (313034666264656338) Stream error: 'Permission denied: Permission denied on resource project roadtrip-royale-dev.'.

That is now working.  Seems the API key changed on me...  Can we now make the Sign Out button work

You are using AuthError.noModelContext, but did not fill it in.

I see 3 issues.  1) we DO NOT ALWAYS (sometimes we do) close the SignInView when the account is created and get logged in (same if we use Sign In). 2) The Google and Apple SignIns don't work.  3) We don't seem to be writing the users to the user database, so someone can sign in with the same name.