Last 2 prompt for User Login (for now)
I want to add a userImage to the User.  This will default to the using the users avatarType and avatarColor to assume there is an image of that name in the assests, ie:"dog_orange" for avatarType of type .dog and avatarColor of type .orange. The user should also be able to select an image and upload it to replace this image. This image should be stored in firebase storage, and cache a local copy (only updating on the next request of that image such as login).

We need to support more data being editable such as the Users first name, last name, and any other important editable data.

FirebaseStorageService.swift does not build.

This is failing to upload the image now, Failed to upload Image: Object <imagepath> does not exist.

Please put back the ImagePickerView, unsure how it disappeared, that was not stated to be removed