Promopt 1 in moreSettings
I have change the branch on the repo to focus on adding more Main Settings for the Main Screen.  Under "User" section, we should ahve an "App Preferences" section.  This should include items like: "Dark Mode" with the options of Light, Dark, System: "Distance Unit" with options of Miles, Kilometers: "Map Style" of Standard, Satellite: "Language" with options of English, Spainish, French. But make it hidden for now: "Play sound effects" which is a toggle. But make it hidden for now: "Use vibrations" which is a Toggle. But make it hidden for now.