Prompt in MoreSetings to make things function.
Wonderful, can we now make the Theme portion work, so the user would use Light Theme if directed, instead of the System settings.  Let also change the map style on all MapViews, so that when Standard is set we use the .hybrid and when Satellite is set we use .standard.