Prompt from MoreSettings branch.
Lets now move onto the "About" section (Help, Support, Contact would be other names you could use).  I want it similar to the attached image but keeping the current theme of course.  We should have the following items: "About" - Clicking on it provides a pop up about HammersTechLLC and the app: "Acknowledgements" - Clicking on it provides a pop up about the SDK's and open source features are using, such as Firebase and Google: "FAQ" - Clicking should bring up an FAQ page about commons questions you can come up with.  Including how the game is played: "Report a Bug" - Sends an email to hammerstechllc@gmail.com, with the headline of "RoadTrip Royale Bug Report": "Suggest a Feature" - Sends an email to hammerstechllc@gmail.com, with the headline of "RoadTrip Royale Feature Suggestion":"Contact Support" - Sends an email to hammerstechllc@gmail.com, with the headline of "RoadTrip Royale Support Issue": We should then have a similar App Version, and then Terms and Privacy that provides a popup with the expected info there.