Prompt in MoreSettings branch to fix overlapping buttons.
I am unable to click on the Terms button only.  It seems no matter what I click on in that VStack (as clicking App Version also does this issue), both showTerms and ShowPrivacy will be set.  So privacy will come up.  Can you please fix this so I can click on them separately.  I even made a background color to show clickable area but it seems to not matter