Prompt in MoreSettings branch. Add privacy/permissions.
Next, I would like to add a Privacy and Permission section. I have attached a sample UI.  I would like this after the User section.  We want the "Location" and to make it clear if that's currently "Allowed" in green or "Disabled" in Red or "While App is Open" in a yellowish color. Clicking on it would bring up the correct location in System Settings.  Below that should be "Save location when marking plates" as a Toggle, "Show my location on large map" as a Toggle, and "Track my location during trips" as a Toggle.  Have a "Microphone" that again makes clear if it's currently allowed by the user in the preferences of the app.  Below that should be one for the "Camera" is allowed or not, but lets hide that for now.  Then a "Notifications" allowed or not.  Below that we should have "Plate found by opponent" as a Toggle, "Plate found by co-pilots" as a Toggle.   "Promotion & News" as a toggle.

Please make sure all these buttons work and save data to AppStorage for now, some may be moved to the User later.

The previous prompt failed at the last attempt as I didn't click Allow to use the cat command, can you finish the previous prompt?