Prompt in MoreSettings branch.  Adding trip settings. Some minor manual UI changes, like making only hte pencil tappable.
Now lets add some Settings to the Trip itself.  At the top should be a section on Trip Info.  All actions should of course update the Trip model.  In this we should first be able to edit the Trip name. Then it should show the createdAt date.  We should have a Start Date that has a "Start Now" button, disabled if not the trip creator. When that button is pressed a startedAt variable is written to, that has a Date. This changes to showing the Date as label once pressed.  This should include a themed Red color "End Trip" button that opens a confirmation dialog informing that this will stop the gamewill write to a variable in the trip model, isTripEnded. This should make it so you can't add states in this game in the UI anymore. Maybe even disabling the Voice tab from working. This End Trip action will also write a tripEndedAt: Date and tripEndedBy: User who ended it in the trip model.
Next would be a reset button.  This will have a confirmation dialog indicated this will reset all trip settings but the trip name. Everything will be reset, including Start Date, which will not auto start.  Any logs would be erased, other than a log stating it was reset. Under would be a Delete button.  Similarly, have a dialog informing this will be deleted and all scores will be removed.

Next we will have Tracking & Privacy. First option would be "Save Location when marking plates".  This would be disabled if permissions are not set for Location.  It would write to a variable in the Trip model for this setting.  Next would be "Show my location on large map", similarly updating the Trip model. Next then "Track my location during trip", again updating the Trip model. Next, "Show my active trip on the large map", similarly update to Trip Model. but would be disabled if "track my location during trip" was set to false.  Finally "Show my active trip on the small map", similarly update to Trip model  but would be disabled if "track my location during trip" was set to false.