Asked some questions about why the code seems to support multiple users
Why would we support more than 1 user opn the ContentView?

Yes please refactor

We don't seem to have a ModelContext ready when we try to create the anonymous user and there is no internet.  How could we fix that, so we initialize with a modelContext even on first run of the app.

So we havea. problem where we create the User, by calling createDefaultUser, when we signInAnonymously(), and add to the modelContext.  But when we get the callback to handleAuthStateChange(), it gives us a firebaseUID and htat doesn't exist on the user yet. SO we make a second user.  How can we fix this?

Getting errors in the Predicate on handleAuthStateChange,  Cannot convert value of type 'PredicateExpressions.Equal<PredicateExpressions.KeyPath<PredicateExpressions.Variable<AppUser>, String>, PredicateExpressions.KeyPath<PredicateExpressions.Value<User>, String>>' to closure result type 'any StandardPredicateExpression<Bool>'