fix/Users Ask cursor about authentication status.
Right now when a user uses an anonymous account, the authentication section still says Signed In, whether I am a Local user (it adds the note "Changes will sync when you're online" or if you are signed in Anonymously.  But that is not correct.  These accounts are NOT signed In.  The log should state soemthing different.  And a Signin/Create account button should be available, instead of a Sign Out button.