Ask seems to be working better at getting me correct work the first time.
I noticed in @FirebaseAuthService.swift in any function, before we try to migrate the user we check something similar to "if let existingUser = currentUser, existingUser.firebaseUID == nil" but since we always have a firebaseUID, even on a Local User.  What would be the best way to determine when to migrate and when to create a new one.