More asking then implementing
When I create an account, it incorrectly doesn't migrate the user.  It seems to be getting caught up in @FirebaseAuthService.swift at func createAccount(..) on this line "guard try await !isUsernameTaken(userName) else {
            throw AuthError.usernameTaken
        }" and is therefore throwing the AuthError.  Should we assume this user is the one who owns it? What are some options here to not immediately throw .usernameTaken error, and instead do more checks to see if it's an anonymous account and if so, then check if this is the same person tied to that anonymous account.  Any other security checks?

Change #3  should not be included in the changes.  It's unneeded, as you are changing updateUserName function.  We already check for the names to not be the same name. That's not used as creation time, but afterwards to change.