fix/Users more asks, and a few prompts.
I have noticed that after I sign in.  If I close the app, and reopen. The app creates an exact copy it would seem of the user to firebase (lastUpdated and all other fields are the same).  It also then thinks this new copy is anonymous, as it marks it as such on the screen, instead of thinking it's logged in.

3 Errors building on the same line: Operator can throw but expression is not marked with 'try', 'try' cannot appear to the right of a non-assignment operator, 'async' call in an autoclosure that does not support concurrency

That didn't fix anything