fix/Users, we are creating a default user different in each path
Is there a reason why createDefaultUser() is not called in loadOrCreateUserFromFirebase?  It's manually created in loadOrCreateUserFromFirebase and how it does it is quite different. Main issue being that it defaults to email address or "User", when we won't have an email address, and we want to use the getDefaultUsername from the DeviceIdentifier.  createDefaultUser() uses localID instead of firebaseUID. And it uses syncLocalUserToFirebase, isntead of saveUserDataToFirebase in loadOrCreateUserFromFirebase.  Are they both correct and we just need to change the username creation, or do they need to be changed?