another ask, but things seem to be reverting in this mixing. I get duplicate and not migrated users still.
Now if I sign out, close the app, and reopen, the app will make a second user with the same username, so it's no longer checking the username when creating a firebase account. Perhaps we should not be reseting the firebaseID, just mark as signed out?  We should also not be allowing duplicate email addresses, can we solve that on the firebase side or do we do that in the app?