Another ask, because things were starting to act weird.
Lets restart this entire User process.  It seems now we are writing to the Users collections, but not making the anonymous user when the app is first opened. This causes a second user to be made, as we are not migrating.   Lets just rewrite the whole @FirebaseAuthService.swift and remove the extra logic.  We want the defaultUser created when the app is opened, using "User(deviceId)(4extraNumbers)".  This will immediately become and anonymous signin if internet is available, if not then we make it local and set it to sync later.  You can then login, and make that account not anonymous by using email/password, Apple login, gmail login, facebook login, instragram login and any toher items allowed by Firebase.

I put in those changes myself, but its not building "Cannot convert value of type "AuthDataResult' to expcted argument type 'User'". This is anywhere we call  "await updateUserFromOAuth("

Great, but we changed @FirebaseAuthService.swift and didn't seem to fix other locations.  For example in UsernameConflictDialog.swift we reference this ".onTapGesture {
                    // Cancel on background tap
                    authService.resolveUsernameConflict(newUsername: nil)
                }
            " but resolveUsernameConflict is gone.  Please fix all parts of the app that were using the old FirebaseAuthService.