Trying to fix the User flow, manully changed sign out to red
Now I want to add lastDateLoggedIn to AppUser, and lastLoginLocation which is an Array of Lat/Long, so we can see the last 5 login locations.  This is optional though, so only works if they enable location. If not enabled, don't ask user, just skip it.

Build error.  Closure captures 'delegate' before it is declared, at let delegate = OneTimeLocationDelegate { [weak self] location in

That didn't fix it, you are trying to capture a object in it's own closure...

Great, but it stated I logged in 3 times quickly.  Are we maybe signing in multiple times?