Prompt to update new trip sheet
Can we please make the Create New Trip sheet be the same style/theme as the rest of the app.  I would also like to make some settings, that if you drag the sheet up become unhidden.  This includes things that are in the Settings of the Trip. But I also want to add a few items.  Checkboxes for including US, Canada and Mexico, defaulting to all 3 checked.  Out trip would then have the same settings so they could be changed later (but only if the trip has not started).  We should also add a checkbox for "Start Trip right away" or "Start trip after creation" or "Start trip". This should also be a default setting in the main settings for a "New Trip Defaults" section.

Error: Referencing instance method 'setValue(forKey:to:)' on 'Array' requires that 'PlateRegion.Country' conform to 'PersistentModel'