More prompts to make the tracker screen show these new settings.
Can you fix the active @TripTrackerView.swift to also have these settings, and to follow the rules (ie: don't show Canada if it's not selected).  I also want you to ONLY be able to click on a state/province or use Voice if the trip is active (started but not stopped).  The Trip option settings though should only be changable once the Trip has been created but not yet started, so started or ended should not allow the change.

Worked great.  But I notice an error in our log.  CoreData: Could not materialize Objective-C class named "Array" from declared attribute value type "Array<String>" of attribute named enabledCountryStrings

Please add the options of United States, Canada and Mexico to the Trip Settings themselves.  As stated before, these should only be allowed BEFORE the trip becomes ACTIVE.  The currently disabled settings for tracking when the Trip is ACTIVE should only be disabled when the trip has Ended, not when it's Active, as I can change that setting while playing without affecting the game

That worked out fine, but can you move the "Countries to inlcude" out of the Tracking and Privacy section into it's own section called something like "Game Settings"