Prompt to change sheet movement
Make the sheet that comes up start by just showing the name, you have to pull the sheet up if you want to access the options.

I want the initial to be even smaller, only show the Basic Info Section, maybe the title of the next section so users know there is more. Keep medium