Prompt to move voice defaults.
The Skip Voice Confirmation default should really be part of the New Trip Defaults.  On that same screen the following changes should be made. The Start trip right away should be the first item like it is in the Game Setting.    the "Default Countries" headline should have some caption text explaining it's use.  All of the Tracking options should not have Dividers between them, and should also have it's own headline about tracking.