Prompts to implement Haptic and Sound
Can we implement these features that are still viisble. Such as haptic feedback and sound feedback

This is failing because the FeedbackService does not conform to ObservableObject protocol.  Please fix

Still not building please build and fix as required.

Cannot find 'appUseVibrations' in scope, is hte error i get.  These variables are in not defined in @ContentView.swift