Prompts to fix the 2nd level Views from popping to root (didn't work well)
If I click anything on the Help and About view that brings up another View, then click the Done button on that View, it brings me back to the Main Settings view, not the Help and About View.

That did not work, we are still jumping back to the Main Settings view.

Lets move the AboutView and other Views in @ContentView.swift that should actually be inside the HelpAboutView.swift file

The DOne button does nothing now.

change the Main Settings to use hte newer NavigationStack.

Add a MainSettingsCoordinator, that uses the Coordinator pattern to make this work.

When we try to click on any of the settings, it moves forward to a blank screen, then auto returns to the Main Settings screen.