Had to manually fix system not going next screen.  It seems I needed to give it the solution.
Tried this last prompt to get it to fix.

I am getting this error when it fails to switch Views some times.

A NavigationLink is presenting a value of type “SettingsDestination” but there is no matching navigationDestination declaration visible from the location of the link. The link cannot be activated.

Note: Links search for destinations in any surrounding NavigationStack, then within the same column of a NavigationSplitView.