More prompts for button clean up.
Perfect, can you make it so the Found and Remaining summaryChips are the same size.  Use whichever is bigger and make them both that size.

I need the summary chips to be the same height and width.  They need to mak sure they do NOT wrap around the text, as it does in this image.

I do not want this to be static.  I want it to be dynamic based on the UI itself

Now make the Start/Stop button the same height as the cards.