iPad onyl prompt for new trip
I have an iPad only issue when we try to add a new trip.  The first image shows right when you click Create new Trip.  This as you can see is not tall enought to see the place to enter text.  Then the second image is the "large" pull.  Can we make that bigger?  If not that would be why 25% doesn't show enough, as 25% is less in tablet.