Ask then prompt to make passwords more secure
Lets change the Password requires to be more secure, this is a family friendly app, so what security do you believe we should implement on the password?

Implement this approach but also require the Upper and Lower case letters. Allows special characters so it can be more secure if an adult wants.