Planned a new feature to sync and download trips.  Storing them in Firebase.  Have not tested yet, but it's building.
PLan: I need to plan a feature for syncing the Trip/Game data with Firebase, so it is remotely accessible from any device the user logs into.  First and foremost, this feature is ONLY for accounts that are NOT Local or Anonymous. They need to provide a full login otherwise the Trips/Games are stored locally on that singular device.  What other settings or features should be implemented as part fo this feature? Second you are not limited to this codebase, if you think I need to make changes on the Firebase side, or create my own server for something, explain it to me. I can immediately think that we need to Sync the data whenever we connect, but it must be in the background unless the data has changed.  But should this only care about changes the current user made, who gets priority or is it all assumed to be valid data?   What if you created a different trip on another device?

Ask: I noticed you didn't check off "Add sync status badges/indicators to trip list and trip detail views". Was that completed?

Ask: Add to both

Plan?: Implemenet the proposed changes

Plan: I am getting a build failure, the Predicate on line 554 of FirebaseTripSyncService.swift has this error: Cannot convert value of type 'PredicateExpressions.Equal<PredicateExpressions.KeyPath<PredicateExpressions.Variable<Trip>, UUID>, PredicateExpressions.KeyPath<PredicateExpressions.Value<Trip>, UUID>>' to closure result type 'any StandardPredicateExpression<Bool>'

Line 125 is complaining: Value of type 'FirebaseAuthService' has no member '$isOnline'.   But I see isOnline as a computed var in FirebaseAuthService

Plan:New error:  @TripTrackerView.swift line 1422, Cannot find 'syncService' in scope.  Seems SettingsView is trying to access a variable that is defined in TripTrackerView.