Plan then prompt switch to Google Maps.
I think it would be a good idea to switch to Google maps instead of Apple Maps.  The main reason being that we can add directions but we can also customize the map look.  The initial custom map should have highlights for each Region. Making it clear where each region starts and ends.  If the region could change color once found that would be better.

Can we implement the Google Maps Migration with Region Boundaries plan?

I addded the SDK and it is building, however I get a startup warning.  Can you check things, as it doesn't see the file, but the file seems to be part of the build.

Google Maps configuration not found. Expected: GoogleMaps-Info-Debug.plist or GoogleMaps-Info.plist
   App will work but maps may not display correctly.