Plan to do minor fixes
I want to change the regions to not be simple block, but be the actual shape of the states themselves.  Can we also make it none transparent, so it looks more solid and playful, think Waze app.  We need to add Canada and Mexico regions as well.  And all other countries should have a single region for them.   finally, when in Satellite or Standard, it should NOT show the custom 2d tiles.

Go ahead and implement that plan