Ask then prompt to fix issues I found.
You added to PlateRegion, Other Countries, but did not exhaust all cases in switch statements that go over it.  Why did you add it to PlateRegion?  I never asked for it to be a plate we can find and add to the list, I only wanted it on the Map itself.

Lets apply these changes but NOT .2 removing WorldCountries.swift as it may be useful in the future to have this list.