PLan and prompt to fix issues
As you can see in the image 3 issues are still happening.  1) There are only US regions visible, no canada, Mexico or rest of world..  2) The regions are still blocks instead of the shape of that region.  3) Even though I have Satellite or Standard selected, the blocks are showing.

imeplement you plan, if I need to get any files inform me.