More plan and fix
That was not what I wanted.  What yoyu changed is showing WHERE the user was  when they found a region.  I want to know what Region was found, similar to what the blocks are doing.  Please do not remove the code I jsut said was wrong, I want to keep that data and be able to show it later.  So just comment that part out.  Instead work the same way we used to, like this screen shots shows.  Green for my current location, with the extra circle.   Blue for not found states, and an orange for found.  I am also including the original TripTrackerView.swift file, since it was using Apple Maps and we can't just copy paste the code.  I want the logic to be the same, just make it work for Google Maps.@/Users/christopherhammers/Downloads/TripTrackerView.swift