More planning to get it looking like it used to
Anything left on that plan to update?

Lets implement the custom green marker for the user location.  Can we also change the colors to match the colors in the old document?  Right now they are purple and a brown orange instead of Blue and Orange.  Can we also make them round, instead of a pin/drop?  Finally, as I scroll through countries I want the mini map to update it's center point like it used to.