More attempts to fix.
That worked great.  2 new issues though: 1) The full screen map view doesn't let me zoom in.  That Screen should all us to zoom in and move, it should not be recentering.  2) The smaller view center points are not working well, they are too zoomed in.  Please use the old values from the previously attached file for the defaults.

You are accessing private coordinator values in the changs to updateUIView