Update TripTrackerView.swift
This did not fix the issue.  It seems that you are doing an onChange(of: enabledCountries) to change what is available.  That should never be done, anywhere.  you then also have cameraPosition being a different variable for RegionMapView and FullScreenMapView. It should instead be a Binding to both of those views from the Main TripTrackerView itself.  By not initializing every refresh, we should be allowed to move.