Plan and Ask, manually test, then finally implemented
You mentioned earlier to get the correct Region outlines :

GeoJSON sources

- Natural Earth Data: https://www.naturalearthdata.com/
- US Census Bureau: https://www.census.gov/geographies/mapping-files.html
- OpenStreetMap: https://www.openstreetmap.org/

Until these files are added, the app will use approximate rectangular boundaries as a fallback. Once added, you'll see the actual region shapes.

How would I go about getting these in a format you need, please give me step by step instructions

I have the files but it seems to not be working.  Says it can't find them, but I did all the checks.  Can you see what might be wrong

OK got it working, as the id was not in the right spot.  But now the shapes it's making are much bigger than the map.  It's unreadable. How can we fix this?

Error on line 260 of this new file, Operator function '!=' requires that 'CLLocationCoordinate2D' conform to 'Equatable'

Neither QGIS or mapshaper.org can change the coordinate system.   Here is an example log from your GeoJSONLoader change:

⚠️ WARNING: Coordinates in mx-states.geojson appear to be in wrong coordinate system!
   Region: mx-hid
   Coordinate range: lon=[5645.0, 5784.0], lat=[5288.0, 5497.0]
   Expected: lon=[-180, 180], lat=[-90, 90] (WGS84 degrees)
   Your file may be in Web Mercator (EPSG:3857) or another projection.
   Please convert to WGS84 (EPSG:4326) geographic coordinates.
   You can use tools like: https://mapshaper.org/ or QGIS to reproject.
⚠️ Skipped all coordinates for mx-hid - wrong coordinate system

The data points for US are in NAD83, how could we convert those, as every conversion just makes the whole image a small item in the US

I got the files to be correct.  But it seems MultiPolygon has issues.  Either not drawing, only drawing 1, or making it a big square.  How can we fix this

Implement the "better" fix (Cursors wording not mine)

Have an error on renderCountryBoundaries.

Cannot convert value of type '[CLLocationCoordinate2D]' to expected argument type 'CLLocationCoordinate2D'