Plan and implement fixes for lag while using boundaries
I want to make a new Plan, Map Lag Fix.  These boundaries have started to make the UI a little laggy.  Is there anyway to reduce the CPU/GPU strain from rendering these boundaries?

It also makes sense that we load the boundaries at start up, instead of the first time we open a TripTrackerView, as it then takes some time to load so there is a noticable delay.

Please include a Splash Screen in this design

Can we implement this plan

@ContentView.swift was incorrectly changed, missing an } to close.  The question becomes was the .transition(.opacity) supposed to be on the ZStack, NavigationStack or AppBackgroundView.