Plan and action for tile redraws
Would we be able to prerender the starndard tile, so we can do it at loading of hte app, or even precompile calculate and save it?  So we only update it when the color changes?

There is no zoom level of 3,  the logs have shown zoom starts at 4 at the farthest out. WIth 21 being the most zoomed in.  We therefore generally work in the 4-8/9 range instead of the 3-5 you suggested.

Yes, implement these changes.  Start with zoom levels 4-7 but make it configurable as you suggested.

The pre-rendered are upside again, but not all of them....Image attached

THE orange redraws are upside....Image attached