More useless asks.
Let's move all the prerendering to #if DEBUG only.   I only see a blank screen instead of the SplashScreenView. That needs to be fixed.

The splash screen doesn't show up even when we had the 280 seconds of preloading.   The fix is not to make it wait 0.1s.  It's just not showing up.

Please imeplement the last 2 prompts but set the delay to 1.0s

Can we fill in a launch screen too