Plan and build
We are very close to done.  Not all items in the Create New Trip screen are translated.  We should also make the selections on any Pickers be localized where possible. Make sure to check all other parts of the app to see where we can localize, since so much has been missed on this first pass.

Final fix, the @TripTrackerView.swift the Section titles of "United States", "Canada", and "Mexico" are not localized.