Found some more items not localized.
I found a few more spots, I need you to check for any others.  When we created a new trip, the detail text of "Automatically start the trip when created" is not localized. "Help & About", it's detail text and a few of the navigated to items are not localized.  The UserProfile screen is also not fully translated.