PLan and build for updating to support both maps after merging wiht main. A few manual changes I found.
We in the Google Maps repo. (the original code was Apple Maps)  Can we make an App Preferences Setting that lets us choice between Apple and Google Maps, that is shown only when it is a DEBUG build.  I have attached the original/current Apple Code files as you have described where changed might need to be.  I want you to combine the files to support both. But we need to rename the GoogleMaps version of the FullScreenMapView and RegionMapView as RegionGoogleMapView, FullScreenGoogleMapView  , and the Apple versions to FullScreeenAppleMapView and RegionAppleMapView

Build that.