Ask and build for a better user ssearch
When searching for Users in the add to family screen, it should show a list of users with that start of the name, then have a checkbox so you can send the invite to multiple users.  The current "Search and Invite User" is odd, because I don't know if found the user.

Build

It doens't seem to be looking at username, as I search UserC and it's not found, but the email they used is.

That is not case insensitive.  Seraching Userc, or userc or USERC does not find it.