Plan and build better search requirements/limitations
Lets add a phone number as part of the search.  But make sure you do NOT include the email or hte phone number in the search if the isEmailPublic and isPhonePublic variables are true for that user.

Build

We are still MATCHING by email.  We don't want to match if isEmailPublic is false.

We are only looking for prefix on the user name, can we instead search the whole username.